+++ feature/auth/src/main/java/com/teamoffroad/feature/auth/data/di/DataModule.kt
@@ -2,8 +2,8 @@ package com.teamoffroad.feature.auth.data.di
 
 import com.teamoffroad.feature.auth.data.datasource.AuthPreferencesDataSource
 import com.teamoffroad.feature.auth.data.datasource.DefaultAuthPreferencesDataSource
-import com.teamoffroad.feature.auth.data.repository.AuthInfoRepository
-import com.teamoffroad.feature.auth.data.repositoryimpl.AuthInfoRepositoryImpl
+import com.teamoffroad.feature.auth.domain.repository.AuthInfoRepository
+import com.teamoffroad.feature.auth.data.repository.AuthInfoRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
